requirements.txt
PySide6>=6.4.0
speedtest-cli>=2.1.3
requests>=2.28.0
pyperclip>=1.8.2
